Fixing case for consistency (#127)
* Fixing case for consistency

Using USERNAME throughout instead of a mix of username and USERNAME to retain consistency.  It also makes it more obvious due to the different case in the URL that it's a user-specific variable.

* Fixing more references

Another username -> USERNAME fix.  Also fixing URL that didn't have the full name of the conflict repo.