Enforce MD029 - Ordered list item prefix
This rule is triggered for ordered lists that do not either start with '1.' or do not have a prefix that increases in numerical order (depending on the configured style). The less-common pattern of using '0.' for all prefixes is also supported.

Rationale: Consistent formatting makes it easier to understand a document.