Enforce MD036 - Emphasis used instead of a heading
This check looks for instances where emphasized (i.e. bold or italic) text is used to separate sections, where a heading should be used instead.

Rationale: Using emphasis instead of a heading prevents tools from inferring the structure of a document.

Note: This rule won't fire if you add punctuation (such as a colon or question mark) to the emphasised text.