Run super linter on `pull_request_target`
This behaves in an almost identical way to the `pull_request` event with the same set of filters and payload. However, instead of running against the workflow and code from the merge commit, the event runs against the workflow and code from the base of the pull request.

This allows the workflow to run on pull requests from forks.